+++ CHANGES.rst
@@ -10,6 +10,9 @@
 - Add support for application-wide callbacks when ``Greenlet`` objects
   are started. See :pr:`1289`, provided by Yury Selivanov.
 
+- It is now possible to consume ready objects using next(gevent.iwait(objs)).
+  Previously such a construction would hang. See :pr:`1288`, provided by Josh
+  Snyder.
 
 1.3.7 (2018-10-12)
 ==================